Created implementation of new invasive predator agent
Invasive predator inherits the predator agent but has a diet that allows it to consume both the prey and the predator agents.

To use, simply set the initial number invasive predators slider off of 0 and hit reset. Purple circle icons should appear representing the invasive predator (will be changed later to a png).